Server Setup: Use curl instead of wget for https scheme
Oracle's VirtualBox, which is the base for Boot2Docker, does not seem to
run correctly on Windows 10, yet. As a consequence, I made
'setup_openfire.sh' runnable on the git bash on Windows 10.

Unfortunately there is no 'wget' delivered with Msysgit. One can download
a 'wget' port that works with MinGW/Msys as described here:
 http://www.kadamwhite.com/archives/2012/wget-in-git-bash
But that one doesn't have https support. So I have to use 'curl' for the
one download that uses https.
I still use wget for the openfire.tgz download. I couldn't make it work
with curl.